sort tests, move HTTPRangeRequestSupportedError, do not override __dir__
